requirements.txt
attrs==18.2.0
audioread==2.1.6
Click==7.0
configparser==3.7.3
cycler==0.10.0
Cython==0.29.1
decorator==4.3.2
dill==0.2.8.2
futures==3.2.0
h5py==2.8.0
joblib==0.12.5
jsonschema==3.0.0
librosa==0.6.2
matplotlib==2.2.3
multiprocess==0.70.6.1
multiprocessing==2.6.2.1
numexpr==2.6.8
numpy==1.15.3
pandas==0.24.1
progress==1.4
python-dateutil==2.8.0
pytools==2016.2.6
PyYAML==3.13
resampy==0.2.1
scikit-learn==0.20.0
scipy==1.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
tables==3.4.4
